added education, experience sections
