+++ frontend/public/165.css
@@ -1,0 +1,4 @@
+.svelte-1jgxwym{margin:0;padding:0}.page-bg.svelte-1jgxwym{background-color:#121212}
+div.skill-buttons.svelte-6y0f3m.svelte-6y0f3m.svelte-6y0f3m{display:flex;flex-flow:row nowrap}div.skill-buttons.svelte-6y0f3m>button.svelte-6y0f3m.svelte-6y0f3m{font-size:1em}div.inputs.svelte-6y0f3m>div.keyword-inputs.svelte-6y0f3m.svelte-6y0f3m{display:flex;flex-flow:column nowrap}div.inputs.svelte-6y0f3m>div.keyword-inputs.svelte-6y0f3m>input.skills-section.svelte-6y0f3m{width:35%;min-width:300px}div.inputs.svelte-6y0f3m>div.keyword-inputs.svelte-6y0f3m>input.skills-keyword.svelte-6y0f3m{width:25%;min-width:280px}
+div.point-section.svelte-l79rkx div.point.svelte-l79rkx{display:flex;flex-flow:row nowrap}div.point-section.svelte-l79rkx div.point input.point-text.svelte-l79rkx{margin:0 0 auto 0}div.point-section.svelte-l79rkx div.point div.point-keywords.svelte-l79rkx{display:flex;flex-flow:column nowrap}
+h1.svelte-1y4xpcp{text-decoration:underline}
