+++ frontend/public/262.262.js
@@ -1,0 +1,1 @@
+(self.webpackChunksvelte_app=self.webpackChunksvelte_app||[]).push([[262,447],{596:(e,t,s)=>{"use strict";s.r(t),s.d(t,{default:()=>g});var n=s(379);function l(e){let t;return{c(){t=(0,n.bG)("div"),t.innerHTML='<h1 class="svelte-1jgxwym">title</h1> \n  <div class="card svelte-1jgxwym"><button class="svelte-1jgxwym">button</button></div>',(0,n.Lj)(t,"class","page-bg svelte-1jgxwym")},m(e,s){(0,n.$T)(e,t,s)},p:n.ZT,i:n.ZT,o:n.ZT,d(e){e&&(0,n.og)(t)}}}class a extends n.f_{constructor(e){super(),(0,n.S1)(this,e,null,l,n.N8,{})}}const r=a;var c=s(231);function u(e){let t,s,l,a,u;return s=new r({}),a=new c.default({}),{c(){t=(0,n.bG)("main"),(0,n.YC)(s.$$.fragment),l=(0,n.Dh)(),(0,n.YC)(a.$$.fragment)},m(e,r){(0,n.$T)(e,t,r),(0,n.ye)(s,t,null),(0,n.$T)(e,l,r),(0,n.ye)(a,e,r),u=!0},p:n.ZT,i(e){u||((0,n.Ui)(s.$$.fragment,e),(0,n.Ui)(a.$$.fragment,e),u=!0)},o(e){(0,n.et)(s.$$.fragment,e),(0,n.et)(a.$$.fragment,e),u=!1},d(e){e&&(0,n.og)(t),(0,n.vp)(s),e&&(0,n.og)(l),(0,n.vp)(a,e)}}}class o extends n.f_{constructor(e){super(),(0,n.S1)(this,e,null,u,n.N8,{})}}const g=o}}]);
