+++ frontend/public/262.css
@@ -1,0 +1,3 @@
+.svelte-1jgxwym{margin:0;padding:0}.page-bg.svelte-1jgxwym{background-color:#121212}
+div.skill-buttons.svelte-6y0f3m.svelte-6y0f3m.svelte-6y0f3m{display:flex;flex-flow:row nowrap}div.skill-buttons.svelte-6y0f3m>button.svelte-6y0f3m.svelte-6y0f3m{font-size:1em}div.inputs.svelte-6y0f3m>div.keyword-inputs.svelte-6y0f3m.svelte-6y0f3m{display:flex;flex-flow:column nowrap}div.inputs.svelte-6y0f3m>div.keyword-inputs.svelte-6y0f3m>input.skills-section.svelte-6y0f3m{width:35%;min-width:300px}div.inputs.svelte-6y0f3m>div.keyword-inputs.svelte-6y0f3m>input.skills-keyword.svelte-6y0f3m{width:25%;min-width:280px}
+h1.svelte-1y4xpcp{text-decoration:underline}
