+++ frontend/src/components/resume_input_form/components/points/points.scss
@@ -1,0 +1,13 @@
+div.point-section{
+    div.point{
+        display: flex;
+        flex-flow: row nowrap;
+        input.point-text{
+            margin: 0 0 auto 0;
+        }
+        div.point-keywords{
+            display: flex;
+            flex-flow: column nowrap;
+        }
+    }
+}
