+++ frontend/src/components/resume_input_form/resume_input_form.scss
@@ -1,0 +1,3 @@
+h1{
+    text-decoration: underline;
+}
