+++ frontend/src/components/resume_input_form/sections/awards/awards.svelte
@@ -1,0 +1,10 @@
+<script>
+    import Section from '../section.svelte';
+    import Points from '../../components/points/points.svelte'; 
+
+    export let awards = [];
+</script>
+
+<Section title='Award'>
+    <Points bind:points={awards}/>
+</Section>
