+++ frontend/src/components/resume_input_form/sections/experience/experience.svelte
@@ -1,6 +1,6 @@
 <script>
-    import Section from './section.svelte';
-    import Points from './points.svelte';
+    import Section from '../section.svelte';
+    import Points from '../../components/points/points.svelte';
     export let title;
     export let workExperiences = [];
     
@@ -33,6 +33,7 @@
                 <input type='date' bind:value={workExperience.endDate}/>
                 <h3 class='location'>Location</h3>
                 <input type='text' bind:value={workExperience.location}/>
+                <h3 class='points'>Job Summary</h3>
                 <Points bind:points={workExperience.points}/>
             </div>
         {/each}
