+++ frontend/src/components/resume_input_form/sections/skills/skills.scss
@@ -1,0 +1,19 @@
+div.skill-buttons{
+    display: flex;
+    flex-flow: row nowrap;
+}
+div.skill-buttons>button{
+    font-size: 1em;
+}
+div.inputs>div.keyword-inputs{
+    display: flex;
+    flex-flow: column nowrap;
+}
+div.inputs>div.keyword-inputs>input.skills-section{
+    width:35%;
+    min-width: 300px;
+}
+div.inputs>div.keyword-inputs>input.skills-keyword{
+    width: 25%;
+    min-width: 280px;
+}
